Changes to the ISSEvt structure to include the ISSCDEvt structure.
A first attempt at including the CdFinder algorithm to handle the reconstruction of fission fragments, but there is a lot of work to be done here to have something physically meaningful.

Also the detector mapping is tweaked to specify each of the rings and sectors separately, introducing the concept of a "side" similar to the DSSSDs of the array.